@@ -2,15 +2,15 @@
     "name": "vue3-moveable",
     "version": "0.24.0",
     "description": "A Vue 3 Component that create Moveable, Draggable, Resizable, Scalable, Rotatable, Warpable, Pinchable, Groupable.",
-    "types": "declaration/index.d.ts",
+    "types": "dist/index.d.ts",
     "main": "dist/moveable.cjs.js",
-    "module": "dist/moveable.esm.js",
+    "module": "dist/moveable.js",
     "sideEffects": false,
+    "type": "module",
     "scripts": {
-        "serve": "vue-cli-service serve demo",
-        "build": "rm -rf dist && rollup -c && npm run declaration",
-        "declaration": "rm -rf declaration && vue-tsc -p tsconfig.declaration.json",
-        "lint": "vue-cli-service lint --no-fix"
+        "dev": "vite",
+        "build": "vite build && vue-tsc",
+        "preview": "vite preview"
     },
     "keywords": [
         "moveable",
@@ -59,14 +59,12 @@
         "declaration/*"
     ],
     "devDependencies": {
-        "@daybrush/builder": "^0.2.4",
-        "@vue/cli-plugin-typescript": "~4.5.0",
-        "@vue/cli-service": "~4.5.0",
-        "@vue/compiler-sfc": "^3.0.0",
-        "rollup-plugin-vue": "^6.0.0",
+        "@vitejs/plugin-vue": "^4.2.3",
         "tslib": "^2.3.1",
-        "typescript": "^4.8",
-        "vue": "^3.1.1",
-        "vue-tsc": "^1.0.7"
+        "typescript": "^4.8 <4.9",
+        "vite": "^4.3.0",
+        "vite-plugin-dts": "^3.0.0-beta.3",
+        "vue": "^3.3.4",
+        "vue-tsc": "^1.8.1"
     }
 }
